[Back]: Add GeoResponse to Weather/Today and Weather/Hourly endpoints

+++ backend/Services/Interfaces/IOpenWeatherService.cs
@@ -4,6 +4,6 @@ namespace Backend.Services.Interfaces;
 
 public interface IOpenWeatherService
 {
-    Task<CurrentWeatherResponse?> GetCurrentAsync(float latitude, float longitude, string token);
-    Task<FiveDayResponse?> GetFiveDayThreeHourAsync(float latitude, float longitude, string token, int count = 40);
+    Task<CurrentWeatherResponse?> GetCurrentAsync(double latitude, double longitude, string token);
+    Task<FiveDayResponse?> GetFiveDayThreeHourAsync(double latitude, double longitude, string token, int count = 40);
 }
